Removing user login, and other changes
The main change proposed here is to remove author field in the Post model,
which doesn't really work at the moment since we don't have login set up.

The second big change is of allowing editing the published_date field,
with the intention of allowing to publish a post in the future.

And the last significant change is a refactoring of the new_post and edit_post
views into a single combined view, since they were almost identical.